+++ chat/server/router.go
@@ -1,4 +1,4 @@
-package chat
+package server
 
 import (
 	"log"
